% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrinio_stock_api.R
\name{intrinio_fetch}
\alias{intrinio_fetch}
\title{Fetch Intrinio Data}
\usage{
intrinio_fetch(url, api_username, api_password)
}
\arguments{
\item{url}{The Intrinio API Endpoint URL to call i.e. api.intrinio.com/prices?ticker=AAPL}

\item{api_username}{Your API username, available at https://intrinio.com/account}

\item{api_password}{Your API password, available at https://intrinio.com/account}
}
\value{
A data frame
}
\description{
This function takes an Intrinio API endpoint URL, API username, and API password as inputs and return a dataframe with the requested data.
Remember to enter your own API username and password from https://intrinio.com/account:
}
\examples{
api_username <- "[YOUR API USERNAME]"
api_password <- "[YOUR API PASSWORD]"
aapl_prices <- intrinio_fetch("api.intrinio.com/prices?ticker=AAPL", api_username, api_password)
}
