% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrinio_stock_api-package.R
\docType{package}
\name{IntrinioStockAPI}
\alias{IntrinioStockAPI}
\alias{IntrinioStockAPI-package}
\title{IntrinioStockAPI - Wrapper for the Intrinio Financial Markets Data and Stock API}
\description{
Download stock prices from the free Intrinio Stock API (<https://intrinio.com/>).
}
\details{
Intrinio offers a REST API which provides financial data 
including stock prices, quotes, technical indicators, company fundamentals, and more. 
Complete documentation for the Intrinio API is available here: 
<https://intrinio.com/documentation/api/>. 
To access the Intrinio API, simply create a free account <https://intrinio.com/>.
}
